// src/app/models/product.ts
export class Product {
    name: string = "";
    description: string = "";
    size: string = "";
    color: string = "";
    price: number = 0;
    slug: string = "";
    imageUrl: string = "";
}


export class NewProduct {
    name: string = "";
    user: string = "";
    description: string = "";
    size: string = "";
    color: string = "";
    price: number = 0;
    slug: string = "";
    imageUrl: string = "";
}